+++ tests/test-chat-template.cpp
@@ -300,13 +300,13 @@ int main(void) {
         }
     };
     std::vector<char> formatted_chat(1024);
-    int32_t res;
+    int64_t res;
 
     // list all supported templates
     std::vector<const char *> supported_tmpl;
     res = llama_chat_builtin_templates(nullptr, 0);
     assert(res > 0);
-    supported_tmpl.resize(res);
+    supported_tmpl.resize((size_t)res);
     res = llama_chat_builtin_templates(supported_tmpl.data(), supported_tmpl.size());
     printf("Built-in chat templates:\n");
     for (auto tmpl : supported_tmpl) {
@@ -329,7 +329,7 @@ int main(void) {
             formatted_chat.data(),
             formatted_chat.size()
         );
-        formatted_chat.resize(res);
+        formatted_chat.resize((size_t)res);
         std::string output(formatted_chat.data(), formatted_chat.size());
         if (output != test_case.expected_output) {
             printf("Expected:\n%s\n", test_case.expected_output.c_str());
